refactor: inline bf-p contract check into JSX conformance runner (#1952)
Move the bf-p children check from a standalone conformance suite into
the existing runJSXConformanceTests flow in jsx-runner.ts. The check
runs after each fixture render and asserts that no bf-p attribute
contains children with HTML markup (scope IDs). This follows the
project convention of a single conformance entry point.
